+++ mne_bids/physio/__init__.py
@@ -1,0 +1,6 @@
+from .eyetracking import (
+    _get_eyetrack_ch_names,
+    _write_eyetrack_tsvs,
+    read_raw_eyetracking_bids
+)
+from ._utils import _get_physio_type
